+++ Changelog.md
@@ -35,9 +35,7 @@ __New Features__
   - Adds heat pump backup sizing methodology input.
   - Add soil and moisture type arguments (for determining ground conductivity and diffusivity) and optional geothermal loop arguments for ground source heat pumps.
   - The "Geometry: Building Number of Units" input is now written to the HPXML `NumberofUnitsInBuilding` element.
-- Manual J design load calculations:
-  - Allow overriding additional design condition inputs: `DailyTemperatureRange` and `...TODO...`.
-  - Miscellaneous improvements.
+- Miscellaneous Manual J improvements.
 - Adds more error-checking for inappropriate inputs (e.g., HVAC SHR=0 or clothes washer IMEF=0).
 - Allow alternative label energy use (W) input for ceiling fans.